requirements.txt
tensorflow-gpu==2.1.0
pandas
numpy
scipy
scikit-learn